feat: add, list, export cmds
All crud operations have been moved to a new db package. We have yet to
extract the export functions and utils. There is also a known bug when
printing the entry indexes (they default to 0).

Basic support for add, list and export were added. More validation is
required for add but its a good start.